# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Marketplace YarmarOK 1.0.0",
    "description": "Web platform where vendors can come together to sell their products or services to customers.\n\nThis is Marketplace YarmarOK Server based on the OpenAPI 3.0 specification\n\nSome useful links:\n- [The YarmarokMarketplace repository](https://github.com/YarmarokMarketplace/marketplace)",
    "contact": {
      "name": "API developers",
      "email": "yarmarokmarketplace@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8081/api",
      "description": "Development Server"
    },
    {
      "url": "https://yarmarok.onrender.com/api",
      "description": "Production Server"
    }
  ],
  "consumes": [
    "application/json",
    "multipart/form-data",
    "text/html"
  ],
  "produces": [
    "application/json",
    "multipart/form-data",
    "text/html"
  ],
...
  "paths": {
...
    "/notices": {
      "get": {
        "summary": "Get all notices",
        "tags": [
          "Notices"
        ],
        "description": "Find all existing notices",
        "operationId": "getAllNotices",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Page number for pagination",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Number of articles per page.",
              "default": 9
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListOfNoticesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "There any notices for this user"
          }
        }
      },
      "post": {
        "summary": "Add new notice",
        "tags": [
          "Notices"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Add new notice",
        "operationId": "addNotice",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/AddNoticeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddNoticeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Not authorized"
          },
          "409": {
            "description": "Email already in use"
          }
        }
      }
    },
    "/notices/{category}": {
      "get": {
        "summary": "Get notices by choosen category",
        "tags": [
          "Notices"
        ],
        "description": "Find and show all existing notices by choosen category",
        "operationId": "getNoticesByCategory",
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "example": "auto",
            "schema": {
              "type": "string",
              "description": "Category for search"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Page number for pagination",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Number of articles per page.",
              "default": 9
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Type of sorting.",
              "enum": [
                "newest",
                "oldest",
                "cheapest",
                "expensive"
              ],
              "default": "newest"
            }
          },
          {
            "name": "goodtype",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Type of filtering.",
              "enum": [
                "new",
                "used"
              ]
            }
          },
          {
            "name": "priceRange",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Type of filtering.",
              "example": "100-199"
            }
          },
          {
            "name": "location",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Type of filtering.",
              "example": "Kyiv"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListOfNoticesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Any notices for the given category"
          }
        }
      }
    },
    "/notices/{id}": {
      "patch": {
        "summary": "Update notice info",
        "tags": [
          "Notices"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Update notice information in the database",
        "operationId": "updateNotice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "example": "6490564f8f85c705511c5de2",
            "schema": {
              "type": "string",
              "description": "Backend-generated unique identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNoticeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddNoticeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Not authorized"
          },
          "404": {
            "description": "Notice not found"
          }
        }
      }
    },
    "/notices/notice/{id}": {
      "delete": {
        "tags": [
          "Notices"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Delete the notice",
        "description": "Delete the notice from the database",
        "operationId": "removeNotice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of notice that needs to be deleted",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Delete the notice with the specified ID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notice deleted"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Not authorized"
          },
          "404": {
            "description": "Notice not found"
          }
        }
      },
      "get": {
        "summary": "Find the notice by ID",
        "tags": [
          "Notices"
        ],
        "description": "Find the notices by it's unique identifier",
        "operationId": "getNoticeById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Find the notices by id",
              "example": "6494c145ddb3a57412f3d4b4"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetNoticeByIdResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/notices/notice/{id}/active": {
      "patch": {
        "summary": "activate and deactivate notice",
        "tags": [
          "Notices"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Update 'active' status in the database and move the notice to the appropriate collection",
        "operationId": "toggleActive",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "example": "6490564f8f85c705511c5de2",
            "schema": {
              "type": "string",
              "description": "Backend-generated unique identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "active": {
                    "type": "boolean",
                    "description": "Status to be changed",
                    "example": false
                  } 
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status is changed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddNoticeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Not authorized"
          },
          "404": {
            "description": "Notice not found"
          }
        }
      }
    },
    "/notices/user/notices": {
      "get": {
        "summary": "Get all user notices",
        "tags": [
          "Notices"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Find all existing notices created by specific user",
        "operationId": "getAllUserNotices",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListOfNoticesResponse"
                }
              }
            }
          },
          "404": {
            "description": "This user has not any notices"
          }
        }
      }
    }, 
    "/notices/favorites/{id}": {
      "post": {
        "summary": "Add the notice to favorites",
        "tags": [
          "Notices"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Add the notice to favorites by it's unique identifier",
        "operationId": "addNoticeToFavorite",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Add the notice by id to favorites"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succesful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddNoticeToFavoriteResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          },
          "404": {
            "description": "'User not found' or 'Notice with {id} not found'"
          }
        }
      },
      "delete": {
        "summary": "Delete the notice from favorites",
        "tags": [
          "Notices"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Delete the notice from favorites by it's unique identifier",
        "operationId": "removeNoticeFromFavorite",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Remove the notice by id to favorites"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notice removed"
          },
          "401": {
            "description": "Not authorized"
          },
          "404": {
            "description": "'User not found' or 'Notice not found in favorite notices' or '${id} invalid format'"
          }
        }
      }
    },
    "/notices/favorites": {
      "get": {
        "summary": "Get all user favorite notices",
        "tags": [
          "Notices"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Find all existing favorites notices created by specific user",
        "operationId": "getFavoriteUserNotices",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListOfNoticesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          },
          "404": {
            "description": "There any notices for this user"
          }
        }
      }
    },
...
    "/notices/search/search-notice": {
      "get": {
        "summary": "Search notice by search query",
        "tags": [
          "Notices"
        ],
        "description": "Find the notices by search query in title and description",
        "operationId": "searchNoticesByKeywords",
        "parameters": [
          {
            "name": "keywords",
            "in": "query",
            "required": true,
            "example": "дитячий велосипед",
            "schema": {
              "type": "string",
              "description": "Keyword for search"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Page number for pagination",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Number of articles per page.",
              "default": 6
            }
          },
          {
            "name": "goodtype",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Type of filtering.",
              "enum": [
                "new",
                "used"
              ]
            }
          },
          {
            "name": "priceRange",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Type of filtering.",
              "example": "100-199"
            }
          },
          {
            "name": "location",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Type of filtering.",
              "example": "Kyiv"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Type of sorting.",
              "enum": [
                "newest",
                "cheapest",
                "expensive"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListOfNoticesResponse"
                }
              }
            }
          },
          "400": {
            "description": "The search query is empty"
          },
          "404": {
            "description": "Notices not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ListOfNoticesResponse": {
        "type": "object",
        "properties": {
          "result": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "isGoodType": {
                  "type": "boolean",
                  "example": true
                },
                "maxPriceInCategory": {
                  "type": "integer",
                  "example": 84450
                },
                "result": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AddNoticeResponse"
                  }
                }
              }
            },
            "description": "List of notices found by request"
          }
        }
      },
      "AddNoticeRequest": {
        "type": "object",
        "required": [
          "title",
          "description",
          "price",
          "location",
          "category"
        ],
        "properties": {
          "title": {
            "type": "string",
            "example": "The chair for sale"
          },
          "description": {
            "type": "string",
            "example": "Very good chair for your home"
          },
          "category": {
            "type": "string",
            "enum": [
              "auto",
              "business-and-services",
              "for-free",
              "children's-world",
              "home-and-garden",
              "help",
              "electricity",
              "spare-parts-for-transport",
              "fashion-and-style",
              "realty",
              "exchange",
              "repair",
              "work",
              "animals",
              "goods-to-win",
              "hobbies-recreation-sports"
            ],
            "example": "home-and-garden"
          },
          "location": {
            "type": "string",
            "example": "Kyiv"
          },
          "price": {
            "type": "integer",
            "example": 2500
          },
          "goodtype": {
            "type": "string",
            "enum": [
              "used",
              "new"
            ],
            "example": "new"
          },
          "photos": {
            "type": "string",
            "format": "binary"
          },
          "contactName": {
            "type": "string",
            "example": "Olena"
          },
          "contactNumber": {
            "type": "string",
            "example": "380961112233"
          }
        }
      },
      "UpdateNoticeRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "The chair for sale"
          },
          "description": {
            "type": "string",
            "example": "Very good chair for your home"
          },
          "category": {
            "type": "string",
            "enum": [
              "auto",
              "business-and-services",
              "for-free",
              "children's-world",
              "home-and-garden",
              "help",
              "electricity",
              "spare-parts-for-transport",
              "fashion-and-style",
              "realty",
              "exchange",
              "repair",
              "work",
              "animals",
              "goods-to-win",
              "hobbies-recreation-sports"
            ],
            "example": "home-and-garden"
          },
          "location": {
            "type": "string",
            "example": "Kyiv"
          },
          "price": {
            "type": "integer",
            "example": 2500
          },
          "comments": {
            "type": "string",
            "example": "Very good chair for you home"
          },
          "goodtype": {
            "type": "string",
            "enum": [
              "used",
              "new"
            ],
            "example": "new"
          },
          "contactName": {
            "type": "string",
            "exapmle": "Olga"
          },
          "contactNumber": {
            "type": "string",
            "exapmle": "380961111111"
          },
          "photos": {
            "type": "string",
            "format": "binary"
          }
        }
      },
...
    }
  },
...
}